Move combine_ansatz_params away from old circuits
